Fix type of parameter used for ArgAction::Count
The type of parameter used for ArgAction::Count should be u8.
Clap will complain as following if used as uszie:

    thread 'main' panicked at
    /home/joukan/.cargo/registry/src/index.crates.io-6f17d22bba15001f/clap_builder-4.5.17/src/parser/error.rs:32:9:

    Mismatch between definition and access of `verbose`. Could not downcast to TypeId { t: 7428646492878894209665195255548636123 },
    need to downcast to TypeId { t: 157166153014986124439104429225962660038 }

+++ bash_readline/bash_readline.rs
@@ -40,7 +40,7 @@ struct Cli {
 
     ///Verbose
     #[clap(short, long, action=clap::ArgAction::Count)]
-    verbose: usize,
+    verbose: u8,
 }
 
 type ReadLine = bash_readline_bss_types::read_line;
